fix: datagrid table accessibility v1
