+++ packages/ibm-products/src/components/Datagrid/Datagrid/DatagridContent.js
@@ -28,7 +28,7 @@ const { TableContainer, Table } = DataTable;
 
 const blockClass = `${pkg.prefix}--datagrid`;
 
-export const DatagridContent = ({ datagridState }) => {
+export const DatagridContent = ({ datagridState, title }) => {
   const { state: inlineEditState, dispatch } = useContext(InlineEditContext);
   const { filterTags, EventEmitter, panelOpen } = useContext(FilterContext);
   const { activeCellId, gridActive, editId } = inlineEditState;
@@ -107,6 +107,7 @@ export const DatagridContent = ({ datagridState }) => {
             ? () => handleGridFocus(inlineEditState, dispatch)
             : null
         }
+        title={title}
       >
         {!withVirtualScroll ? <DatagridHead {...datagridState} /> : null}
         <DatagridBody {...datagridState} rows={rows} />
@@ -245,4 +246,5 @@ DatagridContent.propTypes = {
     getFilterProps: PropTypes.func,
     state: PropTypes.object,
   }),
+  title: PropTypes.string,
 };
